modify sauce function working

+++ backend/controllers/sauce.js
@@ -47,6 +47,10 @@ exports.createSauce = (req, res, next) => {
 exports.deleteSauce = (req, res, next) => {
 	Sauce.findOne({ _id: req.params.id })
 	.then((sauce) => {
+		if(sauce.userId !== req.auth.userId){
+			res.status(403).json({message: "Unauthorised"})
+			return;
+		}
 		const filename = sauce.imageUrl.split("/images/")[1];
 		fs.unlink("images/" + filename, () => {
 			Sauce.deleteOne({ _id: req.params.id })
@@ -96,21 +100,34 @@ exports.likeSauce = (req, res, next) => {
 
 // modify sauce
 exports.modifySauce = (req, res, next) => {
-
-	let sauceObject = {};
-	req.file ? (
-		Sauce.findOne({_id: req.params.id})
-		.then((sauce) => {
-		const filename = sauce.imageUrl.split("/images/")[1];
-		fs.unlink("images/" + filename)}),
-	  sauceObject = {
+	const sauceObject = req.file ? {
 		...JSON.parse(req.body.sauce),
-		imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
-	  }
-	) : ( sauceObject = {...req.body}  )
-	Sauce.updateOne({_id: req.params.id},{...sauceObject,_id: req.params.id})
-	  .then(() => res.status(200).json({message: 'Sauce has been updated'}))
-	  .catch((error) => res.status(400).json({error}))
+		imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+	} : {...req.body};
+  
+	delete sauceObject._userId;
+
+
+	Sauce.findOne({_id: req.params.id})
+		.then((sauce) => {
+			if(sauce.userId !== req.auth.userId){
+				res.status(403).json({message: "Unauthorised"})
+				return;
+			}
+			const filename = sauce.imageUrl.split("/images/")[1];
+			if(req.file){
+                fs.unlink(`images/${filename}`, (error) => {
+                    if(error){
+                        throw error
+                    }
+                })
+            }
+
+			Sauce.updateOne({_id: req.params.id},{...sauceObject,_id: req.params.id})
+			.then(() => res.status(200).json({message: 'Sauce has been updated'}))
+			.catch((error) => res.status(400).json({error}))
+		})
+		.catch((error) => res.status(400).json({error}))
   }
 
 // exports.modifySauce = (req, res, next) => {
